# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coligo Backend API",
        "description": "API documentation for Coligo - Platform for collaborative transport",
        "contact": {
            "email": "support@coligo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/profile/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get user documents",
                "operationId": "getDocuments",
                "responses": {
                    "200": {
                        "description": "Documents retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "documents": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload identity document",
                "operationId": "uploadDocument",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "document_type",
                                    "document_file"
                                ],
                                "properties": {
                                    "document_type": {
                                        "type": "string",
                                        "enum": [
                                            "id_card",
                                            "passport",
                                            "driving_license",
                                            "other"
                                        ]
                                    },
                                    "document_number": {
                                        "type": "string"
                                    },
                                    "document_file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document uploaded successfully"
                    },
                    "409": {
                        "description": "Document type already exists"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}